Rename java packages for lettuce 4 and 5
