+++ dd-java-agent/instrumentation/lettuce-5/src/main/java/datadog/trace/instrumentation/lettuce5/LettuceAsyncCommandsInstrumentation.java
@@ -1,12 +1,8 @@
-package datadog.trace.instrumentation.lettuce;
+package datadog.trace.instrumentation.lettuce5;
 
 import static java.util.Collections.singletonMap;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.isPrivate;
-import static net.bytebuddy.matcher.ElementMatchers.nameEndsWith;
-import static net.bytebuddy.matcher.ElementMatchers.nameStartsWith;
 import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.returns;
 import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 
 import com.google.auto.service.AutoService;
@@ -17,23 +13,23 @@ import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 
 @AutoService(Instrumenter.class)
-public final class LettuceClientInstrumentation extends Instrumenter.Default {
+public class LettuceAsyncCommandsInstrumentation extends Instrumenter.Default {
 
-  public LettuceClientInstrumentation() {
-    super("lettuce");
+  public LettuceAsyncCommandsInstrumentation() {
+    super("lettuce", "lettuce-5-async");
   }
 
   @Override
   public ElementMatcher<TypeDescription> typeMatcher() {
-    return named("io.lettuce.core.RedisClient");
+    return named("io.lettuce.core.AbstractRedisAsyncCommands");
   }
 
   @Override
   public String[] helperClassNames() {
     return new String[] {
       packageName + ".LettuceClientDecorator",
-      packageName + ".LettuceInstrumentationUtil",
-      packageName + ".LettuceAsyncBiFunction"
+      packageName + ".LettuceAsyncBiFunction",
+      packageName + ".LettuceInstrumentationUtil"
     };
   }
 
@@ -41,12 +37,9 @@ public final class LettuceClientInstrumentation extends Instrumenter.Default {
   public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
     return singletonMap(
         isMethod()
-            .and(isPrivate())
-            .and(returns(named("io.lettuce.core.ConnectionFuture")))
-            .and(nameStartsWith("connect"))
-            .and(nameEndsWith("Async"))
-            .and(takesArgument(1, named("io.lettuce.core.RedisURI"))),
+            .and(named("dispatch"))
+            .and(takesArgument(0, named("io.lettuce.core.protocol.RedisCommand"))),
         // Cannot reference class directly here because it would lead to class load failure on Java7
-        packageName + ".ConnectionFutureAdvice");
+        packageName + ".LettuceAsyncCommandsAdvice");
   }
 }
