+++ dd-java-agent/instrumentation/lettuce-5/src/main/java8/datadog/trace/instrumentation/lettuce5/LettuceAsyncCommandsAdvice.java
@@ -1,30 +1,35 @@
-package datadog.trace.instrumentation.lettuce;
+package datadog.trace.instrumentation.lettuce5;
 
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.activateSpan;
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.startSpan;
-import static datadog.trace.instrumentation.lettuce.LettuceClientDecorator.DECORATE;
+import static datadog.trace.instrumentation.lettuce5.LettuceClientDecorator.DECORATE;
+import static datadog.trace.instrumentation.lettuce5.LettuceInstrumentationUtil.expectsResponse;
 
 import datadog.trace.bootstrap.instrumentation.api.AgentScope;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
-import io.lettuce.core.ConnectionFuture;
-import io.lettuce.core.RedisURI;
+import io.lettuce.core.protocol.AsyncCommand;
+import io.lettuce.core.protocol.RedisCommand;
 import net.bytebuddy.asm.Advice;
 
-public class ConnectionFutureAdvice {
+public class LettuceAsyncCommandsAdvice {
 
   @Advice.OnMethodEnter(suppress = Throwable.class)
-  public static AgentScope onEnter(@Advice.Argument(1) final RedisURI redisUri) {
+  public static AgentScope onEnter(@Advice.Argument(0) final RedisCommand command) {
+
     final AgentSpan span = startSpan("redis.query");
     DECORATE.afterStart(span);
-    DECORATE.onConnection(span, redisUri);
+    DECORATE.onCommand(span, command);
+
     return activateSpan(span);
   }
 
   @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
   public static void stopSpan(
+      @Advice.Argument(0) final RedisCommand command,
       @Advice.Enter final AgentScope scope,
       @Advice.Thrown final Throwable throwable,
-      @Advice.Return final ConnectionFuture<?> connectionFuture) {
+      @Advice.Return final AsyncCommand<?, ?, ?> asyncCommand) {
+
     final AgentSpan span = scope.span();
     if (throwable != null) {
       DECORATE.onError(span, throwable);
@@ -33,8 +38,15 @@ public class ConnectionFutureAdvice {
       span.finish();
       return;
     }
-    connectionFuture.handleAsync(new LettuceAsyncBiFunction<>(span));
+
+    // close spans on error or normal completion
+    if (expectsResponse(command)) {
+      asyncCommand.handleAsync(new LettuceAsyncBiFunction<>(span));
+    } else {
+      DECORATE.beforeFinish(span);
+      span.finish();
+    }
     scope.close();
-    // span finished by LettuceAsyncBiFunction
+    // span may be finished by LettuceAsyncBiFunction
   }
 }
