+++ dd-java-agent/instrumentation/lettuce-5/src/main/java8/datadog/trace/instrumentation/lettuce5/LettuceClientDecorator.java
@@ -1,17 +1,14 @@
-package datadog.trace.instrumentation.lettuce;
+package datadog.trace.instrumentation.lettuce5;
 
-import static datadog.trace.instrumentation.lettuce.InstrumentationPoints.getCommandResourceName;
-
-import com.lambdaworks.redis.RedisURI;
-import com.lambdaworks.redis.protocol.RedisCommand;
 import datadog.trace.api.DDSpanTypes;
 import datadog.trace.api.DDTags;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
 import datadog.trace.bootstrap.instrumentation.api.Tags;
 import datadog.trace.bootstrap.instrumentation.decorator.DatabaseClientDecorator;
+import io.lettuce.core.RedisURI;
+import io.lettuce.core.protocol.RedisCommand;
 
 public class LettuceClientDecorator extends DatabaseClientDecorator<RedisURI> {
-
   public static final LettuceClientDecorator DECORATE = new LettuceClientDecorator();
 
   @Override
@@ -40,38 +37,38 @@ public class LettuceClientDecorator extends DatabaseClientDecorator<RedisURI> {
   }
 
   @Override
-  protected String dbUser(RedisURI connection) {
+  protected String dbUser(final RedisURI connection) {
     return null;
   }
 
   @Override
-  protected String dbInstance(RedisURI connection) {
+  protected String dbInstance(final RedisURI connection) {
     return null;
   }
 
   @Override
-  public AgentSpan onConnection(AgentSpan span, RedisURI connection) {
+  public AgentSpan onConnection(final AgentSpan span, final RedisURI connection) {
     if (connection != null) {
       span.setTag(Tags.PEER_HOSTNAME, connection.getHost());
       span.setTag(Tags.PEER_PORT, connection.getPort());
+
       span.setTag("db.redis.dbIndex", connection.getDatabase());
-      span.setTag(DDTags.RESOURCE_NAME, resourceName(connection));
+      span.setTag(
+          DDTags.RESOURCE_NAME,
+          "CONNECT:"
+              + connection.getHost()
+              + ":"
+              + connection.getPort()
+              + "/"
+              + connection.getDatabase());
     }
     return super.onConnection(span, connection);
   }
 
-  public AgentSpan onCommand(AgentSpan span, RedisCommand<?, ?, ?> command) {
-    span.setTag(DDTags.RESOURCE_NAME,
-        null == command ? "Redis Command" : getCommandResourceName(command.getType()));
+  public AgentSpan onCommand(final AgentSpan span, final RedisCommand command) {
+    final String commandName = LettuceInstrumentationUtil.getCommandName(command);
+    span.setTag(
+        DDTags.RESOURCE_NAME, LettuceInstrumentationUtil.getCommandResourceName(commandName));
     return span;
-  }
-
-  private static String resourceName(RedisURI connection) {
-    return "CONNECT:"
-      + connection.getHost()
-      + ":"
-      + connection.getPort()
-      + "/"
-      + connection.getDatabase();
   }
 }
